+++ projects/openapi-utilities/src/index.ts
@@ -48,6 +48,12 @@ export * from './ci-types';
 export { generateSpecResults } from './openapi3/implementations/openapi3/generate-spec-results';
 export * from './openapi3/implementations/openapi3/types';
 
+export {
+  getFactForJsonPath,
+  constructFactTree,
+  FactTree,
+} from './openapi3/json-path-utilities';
+
 export { UserError } from './errors';
 export {
   findOpticCommentId,
