+++ projects/rulesets-base/src/extended-rules/spectral-rule.ts
@@ -1,9 +1,15 @@
 import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
-import { IFact, IChange, OpenAPIV3, Result } from '@useoptic/openapi-utilities';
-import { ExternalRuleBase } from './external-rule-base';
+import {
+  IFact,
+  IChange,
+  OpenAPIV3,
+  Result,
+  constructFactTree,
+  getFactForJsonPath,
+} from '@useoptic/openapi-utilities';
+import { ExternalRuleBase } from '../rules/external-rule-base';
 
 type Lifecycle = 'added' | 'addedOrChanged' | 'changed' | 'always';
-type FactTree = Record<string, { fact: IFact | null; children: FactTree }>;
 
 // Spectral results will be returned if they fail
 export interface SpectralResult {
@@ -32,44 +38,6 @@ function toOpticResult(
   };
 }
 
-function constructFactTree(facts: IFact[]): FactTree {
-  const factTree: FactTree = {};
-  for (const fact of facts) {
-    let tree: FactTree = factTree;
-    const pathParts = jsonPointerHelpers.decode(fact.location.jsonPath);
-    const leadingPaths = pathParts.slice(0, -1);
-    const trailingPath = pathParts.slice(-1)[0];
-
-    for (const part of leadingPaths) {
-      if (!tree[part]) {
-        tree[part] = { fact: null, children: {} };
-      } else {
-        tree = tree[part].children;
-      }
-    }
-    tree[trailingPath].fact = fact;
-  }
-
-  return factTree;
-}
-
-function getFactForSpectralResult(
-  spectralResult: SpectralResult,
-  factTree: FactTree
-): IFact | null {
-  let fact: IFact | null = null;
-  const pathParts = spectralResult.path.map((p) => String(p));
-  let tree = factTree;
-  for (const part of pathParts) {
-    const node = tree[part];
-    if (node.fact) {
-      fact = node.fact;
-    }
-    tree = node.children;
-  }
-  return fact;
-}
-
 export class SpectralRule extends ExternalRuleBase {
   private lifecycle: Lifecycle;
   constructor(
@@ -90,7 +58,7 @@ export class SpectralRule extends ExternalRuleBase {
     nextJsonLike: OpenAPIV3.Document<{}>;
     currentJsonLike: OpenAPIV3.Document<{}>;
   }): Promise<Result[]> {
-    const factTree: FactTree = constructFactTree(inputs.nextFacts);
+    const factTree = constructFactTree(inputs.nextFacts);
 
     const changesByJsonPath: Record<string, IChange> = inputs.changelog.reduce(
       (acc, next) => {
@@ -104,9 +72,11 @@ export class SpectralRule extends ExternalRuleBase {
     const results: Result[] = [];
 
     for (const spectralResult of spectralResults) {
-      const fact = getFactForSpectralResult(spectralResult, factTree);
+      const path = jsonPointerHelpers.compile(
+        spectralResult.path.map((p) => String(p))
+      );
+      const fact = getFactForJsonPath(path, factTree);
       if (!fact) {
-        // TODO figure out if this is possible
         continue;
       }
